Add my solution for day 7.
For this solution, I went with a full filesystem model, and the very
basics of a command model approach (though the latter is very basic and
tailored for these needs).

The filesystem model consists of a `BaseNode` class, which represents a
node in a filesystem. `DirNode` subclasses that to track children and to
compute its size from the sum of the childrens' sizes (recursively if
needed). `FileNode` tracks a file with a set size.

The command parser stores the current command line, then grabs all
input, and finally feeds it all back into a command handler. The commadn
handler can then analyze that to figure out what filesystem operations
to conduct behind the scenes. (This feels a little backwards from
implementing actual command line support, since the command handlers are
parsing results, not emitting them).

I went overboard with the filesystem operations. I have path
normalization built-in (allowing for crazy things like
`/foo/../bar/.//foobar`), and let commands like `cd` take such paths.
Might be useful if we ever circle back to this task.

More details are in `task1.py` and `task2.py`.