Add my solutions for day 8.
Like with my previous solutions, I aimed for runtime performance and
memory here.

I do need the whole map in memory in order to actually compute the
visibility and optimal tree spots. However, I can optimize runtime
performance.

To do that, I reduce the number of iterations required. I perform
horizontal scans of the tree as I load in lines from the dataset,
avoiding a second loop there.

When I do the vertical scans, I perform a top-to-bottom scan, followed
by a bottom-to-top scan. In the bottom-to-top scan, I perform the
calculations necessary to get the result, since everything needed for
the results are now in the tree. This saves me another pass through the
grid of trees.

There's a lot more to this. Details are in `task1.py` and `task2.py`.