Add my solutions for day 10.
This was my birthday, and I didn't get to all of it right away.

This solution involves a simple virtual machine, executing a simple
instruction set and keeping track of a single X register.

task1.py calculates a "signal strength" periodically based on X.

task2.py draws a "sprite" to a "screen".

Details are in comments in the code.