Add my solutions for Day 11.
Day 11 involves parsing monkeys that like to throw things.

task1.py was pretty easy. I wrote more code than usual, to model
monkeys, operations, expressions, and commands.

task2.py has a tricky part where the math causes numbers to grow to
crazy levels. There are a lot of approaches to this. I think mine is the
most reasonable, given the task.

Details are in the comments.