Add my solutions for day 14.
This one is all about having a sandy good time.

Details on my approach are in task1.py and task2.py. A notable thing is
that I don't bother with using a grid for the map, as that's wasteful.
Instead, I use a hashtable to populate only the data I need to test
against.