Add my solutions for day 16.
This one was a challenge. My graph theory is not strong, but I
eventually worked my way through it.

Task 2 ended up being a fairly trivial addition onto Task 1, but has the
odd quirk of succeeding for the real input, but failing for the test
input. So my approach may be wrong, but I have to move on from this one.