Add my solutions for day 20.
This is a pretty simple task, but one I spent far too long on. We have a
list of numbers, and have to shift each number left or right in the list
by the value of that number.

I worked for a while on more efficient, optimal approaches, so I
wouldn't have to just use a Python list (which is inefficient for these
operations).

In the end, I used a Python list, because it's still plenty fast enough,
and this isn't production code that needs to be performant.