Add my solution for day 1, 2023.
It's Advent of Code time again! In this one, we're working on scanning
through lines of input, looking for the first and last digits, squashing
them together into a new number, and summing the total results of each
line.

In task 1, we're happily dealing solely with 0..9 values.

In task 2, we have to deal with words representing numbers, like "one",
"two", "three", etc. This is a bit more complex, particularly as when we
walk backwards, we can end up with some overlap between words.

My approaches are documented in each task file.