Add my solution for 2024, Day 1.
This was a simple starter. In both tasks, we have two lists of location
IDs (integers).

In task 1, we need to compare items in the lists from lowest to highest,
calculate deltas, and sum them.

In task 2, we need to track how many times items appear in list 2, and
then sum the results of `location_id * hit_count` based on location IDs
in list 1.

I'm aiming for *reasonably* optimal solutions this year that are
reasonable. There are ways to optimize this for lowest memory usage,
lowest runtime, fewest operations performed, etc. My goal is to make
code that is readable but doesn't take a long time to run.