Add my solution for 2024, Day 2.
In both tasks, we're working with "reports" (lines of integers) and
checking for errors. An error is a number that isn't a delta between 1
and 3 from its neighbors, or a line that isn't all ascending or
descending numbers.

In task 1, we had to count how many reports were invalid.

In task 2, we did the same, but had to allow for up to one error per
report. This required more fault tolerance and some variations on the
type of error that could happen.