Add my solution for 2024, Day 3.
We're already getting into the "let's simulate a CPU!" portion of this
year's Advent of Code.

In both tasks, we're dealing with corrupted code we want to try
executing anyway. What a good idea.

In task 1, we just care about `mul(X,Y)` instructions. Easily solved by
looking for complete regexes.

In task 2, we have more instructions to match, and state to manage.
Seems the right time to start defining instructions more formally.