Add my solution for 2024, Day 4.
Today's tasks are word search problems.

In Task 1, we're looking for any "XMAS" strings and counting then. These
may appear in any direction, and any position within the input may have
more than one of these. We just search through, figure out search
deltas, and look for the full strings. Pretty easy.

In Task 2, we're looking for X-shapeod "MAS" strings. This is actually a
little bit easier. We just need to find any position with an "A", then
compute the adjacent corners and check them for the "M" and "S".