Add my solution for 2024, Day 5.
Today's problem is about building sort functions. The problem involves
checking lists of page numbers and seeing if they're in the right order.
That order is defined by mappings showing which page numbers are before
which other page numbers.

I honestly thought this would involve more factors to check (building a
mapping and a reverse mapping of these page numbers, building a graph,
something).

In Task 1, we're searching for lists of page numbers that are already in
the right order.

In Task 2, we're searching for ones not in the right order.

The difference between the tasks is a single character. A comparison of
`==` vs `!=`.