Add my solution for 2024, Day 6.
Today's tasks deal with movement along a 2D grid, with obstacles that
cause the character (a guard) to rotate clockwise 90 degrees whenever
hit. This requires parsing map information, implementing walking, and
then accomplishing tasks.

In task 1, we're simply count how many spaces we walk before walking off
the board.

In task 2, we have a more complex goal. We need to determine locations
for obstacles we can put in the guard's way that will cause the guard to
infinite-loop. While I'm sure there are some clever approaches to
minimize the time required (I've implemented a couple, but not all ideas
panned out), the approach I ended up with is just to place an obstacle
in front of the guard on every step and then do a "have I been here
before?" loop detection check for each of those. This takes a handful of
seconds, so I'm sure there's a more optimal solution, but I have to move
on.