Add my solution for 2024, Day 7.
Today's tasks are all about missing math operators. We're given an
answer and some values going into an equation. We have to find the
operators. Not all equations will have a valid solution.

In task 1, we only have two operators to try in-between values: Addition
and Multiplication.

In task 2, we also have a "concat" operator, where we concatenate two
numbers together. Pretty easy in Python.