Add my solution for 2024, Day 8.
These tasks are all about scanning a map for antennas of a particular
type, and then figuring out where to create antinodes on the map.

In task 1, we just need to figure out where to place antinodes for each
pair of antennas.

In task 2, we do the same but we repeat at equals along the line until
we're off the map.